requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic-settings==2.1.0
xgboost==2.0.3
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
joblib==1.3.2
python-multipart
pytest>=7.0.0
httpx>=0.24.0
